Add jwt and user_id into task: delete, patch, get

+++ routes/tasks/task.delete.js
@@ -2,6 +2,7 @@ const router = require('express').Router();
 const { query } = require('express-validator');
 const db = require('../../models/index.js');
 const checkError = require('../../checkErrorValidation/checkError.js');
+const jwt = require('jsonwebtoken');
 
 router.route('/tasks').delete(
         query('uuid').isUUID(),
@@ -9,7 +10,9 @@ router.route('/tasks').delete(
     async (req, res) => {
         try {
             const { uuid } = req.query;
-            await db.Task.destroy({ where: { uuid } });
+            const decoded = jwt.verify(req.headers.authorization, 'chereshnia');
+            const user_id = decoded.id;
+            await db.Task.destroy({ where: { uuid, user_id } });
             res.sendStatus(202);
         } catch(err) {
             res.status(400).send(err);
